HSEARCH-3269 Move and adapt HS5's jsr352 jberet module
