+++ batch-jsr352/jberet/src/main/resources/META-INF/beans.xml
