+++ batch-jsr352/jberet/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,0 +1,1 @@
+org.hibernate.search.batch.jsr352.jberet.impl.HibernateSearchJsr352Extension
