+++ batch-jsr352/jberet/src/main/resources/META-INF/services/org.jberet.spi.JobXmlResolver
@@ -1,0 +1,1 @@
+org.hibernate.search.batch.jsr352.jberet.impl.HibernateSearchJobXmlResolver
