--- legacy/jsr352/jberet/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,0 @@
-org.hibernate.search.jsr352.jberet.impl.HibernateSearchJsr352Extension
