--- legacy/jsr352/jberet/src/main/resources/META-INF/services/org.jberet.spi.JobXmlResolver
@@ -1,1 +1,0 @@
-org.hibernate.search.jsr352.jberet.impl.HibernateSearchJobXmlResolver
