--- legacy/jsr352/pom.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
- ~ Hibernate Search, full-text search for your domain model
- ~
- ~ License: GNU Lesser General Public License (LGPL), version 2.1 or later
- ~ See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
-  -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <artifactId>hibernate-search-legacy-parent</artifactId>
-        <groupId>org.hibernate</groupId>
-        <version>6.0.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>hibernate-search-jsr352-parent</artifactId>
-    <packaging>pom</packaging>
-
-    <name>Hibernate Search JSR-352 Aggregator</name>
-    <description>Hibernate Search JSR-352 integration aggregator POM</description>
-
-    <modules>
-        <module>core</module>
-        <module>jberet</module>
-    </modules>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-checkstyle-plugin</artifactId>
-            </plugin>
-        </plugins>
-    </build>
-
-</project>
